[NEXT-FE-9] ensure metadata
